feat: add getIdentity method to the router to provide ER proximity info
This PR add support for solana RPC getIdentity method, which is used to
query the closest ER node to the router (ping time wise). For every
upstream node, router maintains a proximity metric, measureed in ping
latency. This metric might not be very accurate if the ER is behind some
proxy or if the target host simply doesn't handle ICMP packets.